+++ workspaces/quickstart/plugins/quickstart/src/index.ts
@@ -26,8 +26,24 @@ export * from './plugin';
 
 export { useQuickstartDrawerContext } from './hooks/useQuickstartDrawerContext';
 export type { QuickstartDrawerContextType } from './components/QuickstartDrawerContext';
+/**
+ * @public
+ */
+export type {
+  UserRole,
+  QuickstartItemData,
+  QuickstartItemCtaData,
+} from './types';
+export { QuickstartButton } from './components/QuickstartButton/QuickstartButton';
+export type { QuickstartButtonProps } from './components/QuickstartButton/QuickstartButton';
+/**
+ * @public
+ */
+export { QuickstartDrawerProvider } from './components/QuickstartDrawerProvider';
 
 export {
   quickstartTranslations,
   quickstartTranslationRef,
 } from './translations';
+
+export { filterQuickstartItemsByRole } from './utils';